# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Digilist Backoffice API
  description: |
    Complete API for the Digilist facility booking and management platform.

    ## Authentication

    Most endpoints require authentication via JWT token in the `Authorization` header:
    ```
    Authorization: Bearer <jwt-token>
    ```

    Additional headers for tenant context:
    - `X-Tenant-Id`: UUID of the tenant (required for most endpoints)
    - `X-License-Key`: License key for API access

    ## Rate Limiting

    API requests are limited to 1000 requests per minute per tenant.

    ## Response Format

    All responses follow a standard format:

    **Success (single resource):**
    ```json
    { "data": { ... } }
    ```

    **Success (list):**
    ```json
    { "data": [...], "meta": { "total": 100, "page": 1, "limit": 20, "totalPages": 5 } }
    ```

    **Error:**
    ```json
    { "error": { "code": "ERROR_CODE", "message": "Human readable message" } }
    ```
  version: 1.0.0
  contact:
    name: Xala Technologies
    email: api@xalatechnologies.com
    url: https://xalatechnologies.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.digilist.no
    description: Production server
  - url: http://localhost:4000
    description: Local development
...
paths:
...
  /api/settings:
    get:
      tags: [Settings]
      summary: Get tenant settings
      operationId: getTenantSettings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: Tenant settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TenantSettings"

    put:
      tags: [Settings]
      summary: Update tenant settings
      operationId: updateTenantSettings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/TenantSettings"
  /api/settings/integrations:
    get:
      tags: [Settings]
      summary: Get integration settings
      operationId: getIntegrationSettings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
      responses:
        "200":
          description: Integration settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/IntegrationSettings"
  /api/settings/integrations/{provider}:
    put:
      tags: [Settings]
      summary: Update integration settings
      operationId: updateIntegrationSettings
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/TenantIdHeader"
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [bankid, vipps, visma, rco, brreg, outlook, googleCalendar]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated integration settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/IntegrationSettings"

  # ==========================================================================
  # Discount Codes Endpoints
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    TenantIdHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    TenantSettings:
      type: object
      properties:
        tenantId:
          type: string
        displayName:
          type: string
        logo:
          type: string
        primaryColor:
          type: string
        timezone:
          type: string
          default: Europe/Oslo
        currency:
          type: string
          default: NOK
        language:
          type: string
          default: "no"
    IntegrationSettings:
      type: object
      properties:
        bankid:
          type: object
          properties:
            enabled:
              type: boolean
        vipps:
          type: object
          properties:
            enabled:
              type: boolean
            merchantId:
              type: string
        visma:
          type: object
          properties:
            enabled:
              type: boolean
            companyId:
              type: string
        rco:
          type: object
          properties:
            enabled:
              type: boolean
            apiKey:
              type: string
...
security:
  - bearerAuth: []